feat: move `async_db_proxy` fixture to root `conftest`
