+++ tests/tasks/users/conftest.py
@@ -15,30 +15,10 @@
 from typing import TYPE_CHECKING
 
 import pytest
-from sqlalchemy.ext.asyncio import AsyncSession
 
 if TYPE_CHECKING:
     from pytest_mock import MockerFixture
-    from sqlalchemy.orm import Session
-
-
-@pytest.fixture
-def async_db_proxy(mocker: "MockerFixture", sync_db: "Session") -> "AsyncSession":
-    """Create a mocked `AsyncSession` that proxies to the sync session. This will allow us to execute the async CLI commands
-    and then in the unit test function use the sync session to assert the changes.
-
-    Args:
-        mocker: pytest-mock fixture.
-        sync_db: Sync session.
-
-    Returns:
-        Mocked `AsyncSession` that proxies to the sync session.
-    """
-    async_session = AsyncSession()
-    async_session.sync_session = sync_db
-    async_session._proxied = sync_db
-    async_session.close = mocker.AsyncMock()
-    return async_session
+    from sqlalchemy.ext.asyncio import AsyncSession
 
 
 @pytest.fixture(autouse=True)
